Add make integration-actions command
