+++ integration/up_test.go
@@ -57,14 +57,8 @@ var (
 	deploymentTemplate        = template.Must(template.New("deployment").Parse(deploymentFormat))
 	statefulsetTemplate       = template.Must(template.New("statefulset").Parse(statefulsetFormat))
 	manifestTemplate          = template.Must(template.New("manifest").Parse(manifestFormat))
-	user                      = ""
-	kubectlBinary             = "kubectl"
 	zero                int64 = 0
 )
-
-type deployment struct {
-	Name string
-}
 
 const (
 	indexEndpoint    = "http://localhost:8080/index.html"
@@ -176,21 +170,6 @@ persistentVolume:
 )
 
 var mode string
-
-func TestMain(m *testing.M) {
-	if u, ok := os.LookupEnv("OKTETO_USER"); !ok {
-		log.Println("OKTETO_USER is not defined")
-		os.Exit(1)
-	} else {
-		user = u
-	}
-
-	if runtime.GOOS == "windows" {
-		kubectlBinary = "kubectl.exe"
-	}
-
-	os.Exit(m.Run())
-}
 
 func TestGetVersion(t *testing.T) {
 	v, err := utils.GetLatestVersionFromGithub()
@@ -1159,19 +1138,6 @@ func deploy(ctx context.Context, namespace, name, path string, isDeployment bool
 	return nil
 }
 
-func writeDeployment(template *template.Template, name, path string) error {
-	dFile, err := os.Create(path)
-	if err != nil {
-		return err
-	}
-
-	if err := template.Execute(dFile, deployment{Name: name}); err != nil {
-		return err
-	}
-	defer dFile.Close()
-	return nil
-}
-
 func writeStatefulset(name, path string) error {
 	dFile, err := os.Create(path)
 	if err != nil {
@@ -1183,41 +1149,6 @@ func writeStatefulset(name, path string) error {
 	}
 	defer dFile.Close()
 	return nil
-}
-
-func getOktetoPath(ctx context.Context) (string, error) {
-	oktetoPath, ok := os.LookupEnv("OKTETO_PATH")
-	if !ok {
-		oktetoPath = "/usr/local/bin/okteto"
-	}
-
-	log.Printf("using %s", oktetoPath)
-
-	var err error
-	oktetoPath, err = filepath.Abs(oktetoPath)
-	if err != nil {
-		return "", err
-	}
-
-	cmd := exec.Command(oktetoPath, "version")
-	cmd.Env = os.Environ()
-
-	o, err := cmd.CombinedOutput()
-	if err != nil {
-		return "", fmt.Errorf("okteto version failed: %s - %s", string(o), err)
-	}
-
-	log.Println(string(o))
-	return oktetoPath, nil
-}
-
-func getDeployment(ctx context.Context, ns, name string) (*appsv1.Deployment, error) {
-	client, _, err := K8sClient()
-	if err != nil {
-		return nil, err
-	}
-
-	return client.AppsV1().Deployments(ns).Get(ctx, name, metav1.GetOptions{})
 }
 
 func getStatefulset(ctx context.Context, ns, name string) (*appsv1.StatefulSet, error) {
